+++ tencentcloud/dc/v20180410/models.go
@@ -912,6 +912,22 @@ type DirectConnect struct {
 	// 物理专线是否已签署用户协议
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	SignLaw *bool `json:"SignLaw,omitempty" name:"SignLaw"`
+
+	// 物理专线是否为LocalZone
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	LocalZone *bool `json:"LocalZone,omitempty" name:"LocalZone"`
+
+	// 该物理专线下vlan 0的专线通道数量
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	VlanZeroDirectConnectTunnelCount *uint64 `json:"VlanZeroDirectConnectTunnelCount,omitempty" name:"VlanZeroDirectConnectTunnelCount"`
+
+	// 该物理专线下非vlan 0的专线通道数量
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	OtherVlanDirectConnectTunnelCount *uint64 `json:"OtherVlanDirectConnectTunnelCount,omitempty" name:"OtherVlanDirectConnectTunnelCount"`
+
+	// 物理专线最小带宽
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	MinBandwidth *uint64 `json:"MinBandwidth,omitempty" name:"MinBandwidth"`
 }
 
 type DirectConnectTunnel struct {