+++ tencentcloud/tcr/v20190924/client.go
@@ -393,6 +393,56 @@ func (c *Client) CreateSecurityPolicy(request *CreateSecurityPolicyRequest) (res
     return
 }
 
+func NewCreateTagRetentionExecutionRequest() (request *CreateTagRetentionExecutionRequest) {
+    request = &CreateTagRetentionExecutionRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "CreateTagRetentionExecution")
+    return
+}
+
+func NewCreateTagRetentionExecutionResponse() (response *CreateTagRetentionExecutionResponse) {
+    response = &CreateTagRetentionExecutionResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 手动执行版本保留
+func (c *Client) CreateTagRetentionExecution(request *CreateTagRetentionExecutionRequest) (response *CreateTagRetentionExecutionResponse, err error) {
+    if request == nil {
+        request = NewCreateTagRetentionExecutionRequest()
+    }
+    response = NewCreateTagRetentionExecutionResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewCreateTagRetentionRuleRequest() (request *CreateTagRetentionRuleRequest) {
+    request = &CreateTagRetentionRuleRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "CreateTagRetentionRule")
+    return
+}
+
+func NewCreateTagRetentionRuleResponse() (response *CreateTagRetentionRuleResponse) {
+    response = &CreateTagRetentionRuleResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 创建版本保留规则
+func (c *Client) CreateTagRetentionRule(request *CreateTagRetentionRuleRequest) (response *CreateTagRetentionRuleResponse, err error) {
+    if request == nil {
+        request = NewCreateTagRetentionRuleRequest()
+    }
+    response = NewCreateTagRetentionRuleResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewCreateUserPersonalRequest() (request *CreateUserPersonalRequest) {
     request = &CreateUserPersonalRequest{
         BaseRequest: &tchttp.BaseRequest{},
@@ -764,6 +814,31 @@ func (c *Client) DeleteSecurityPolicy(request *DeleteSecurityPolicyRequest) (res
         request = NewDeleteSecurityPolicyRequest()
     }
     response = NewDeleteSecurityPolicyResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewDeleteTagRetentionRuleRequest() (request *DeleteTagRetentionRuleRequest) {
+    request = &DeleteTagRetentionRuleRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "DeleteTagRetentionRule")
+    return
+}
+
+func NewDeleteTagRetentionRuleResponse() (response *DeleteTagRetentionRuleResponse) {
+    response = &DeleteTagRetentionRuleResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 删除版本保留规则
+func (c *Client) DeleteTagRetentionRule(request *DeleteTagRetentionRuleRequest) (response *DeleteTagRetentionRuleResponse, err error) {
+    if request == nil {
+        request = NewDeleteTagRetentionRuleRequest()
+    }
+    response = NewDeleteTagRetentionRuleResponse()
     err = c.Send(request, response)
     return
 }
@@ -1393,6 +1468,81 @@ func (c *Client) DescribeSecurityPolicies(request *DescribeSecurityPoliciesReque
     return
 }
 
+func NewDescribeTagRetentionExecutionRequest() (request *DescribeTagRetentionExecutionRequest) {
+    request = &DescribeTagRetentionExecutionRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "DescribeTagRetentionExecution")
+    return
+}
+
+func NewDescribeTagRetentionExecutionResponse() (response *DescribeTagRetentionExecutionResponse) {
+    response = &DescribeTagRetentionExecutionResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 查询版本保留执行记录
+func (c *Client) DescribeTagRetentionExecution(request *DescribeTagRetentionExecutionRequest) (response *DescribeTagRetentionExecutionResponse, err error) {
+    if request == nil {
+        request = NewDescribeTagRetentionExecutionRequest()
+    }
+    response = NewDescribeTagRetentionExecutionResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewDescribeTagRetentionExecutionTaskRequest() (request *DescribeTagRetentionExecutionTaskRequest) {
+    request = &DescribeTagRetentionExecutionTaskRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "DescribeTagRetentionExecutionTask")
+    return
+}
+
+func NewDescribeTagRetentionExecutionTaskResponse() (response *DescribeTagRetentionExecutionTaskResponse) {
+    response = &DescribeTagRetentionExecutionTaskResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 查询版本保留执行任务
+func (c *Client) DescribeTagRetentionExecutionTask(request *DescribeTagRetentionExecutionTaskRequest) (response *DescribeTagRetentionExecutionTaskResponse, err error) {
+    if request == nil {
+        request = NewDescribeTagRetentionExecutionTaskRequest()
+    }
+    response = NewDescribeTagRetentionExecutionTaskResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewDescribeTagRetentionRulesRequest() (request *DescribeTagRetentionRulesRequest) {
+    request = &DescribeTagRetentionRulesRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "DescribeTagRetentionRules")
+    return
+}
+
+func NewDescribeTagRetentionRulesResponse() (response *DescribeTagRetentionRulesResponse) {
+    response = &DescribeTagRetentionRulesResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 查询版本保留规则
+func (c *Client) DescribeTagRetentionRules(request *DescribeTagRetentionRulesRequest) (response *DescribeTagRetentionRulesResponse, err error) {
+    if request == nil {
+        request = NewDescribeTagRetentionRulesRequest()
+    }
+    response = NewDescribeTagRetentionRulesResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewDescribeUserQuotaPersonalRequest() (request *DescribeUserQuotaPersonalRequest) {
     request = &DescribeUserQuotaPersonalRequest{
         BaseRequest: &tchttp.BaseRequest{},
@@ -1789,6 +1939,31 @@ func (c *Client) ModifySecurityPolicy(request *ModifySecurityPolicyRequest) (res
         request = NewModifySecurityPolicyRequest()
     }
     response = NewModifySecurityPolicyResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewModifyTagRetentionRuleRequest() (request *ModifyTagRetentionRuleRequest) {
+    request = &ModifyTagRetentionRuleRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("tcr", APIVersion, "ModifyTagRetentionRule")
+    return
+}
+
+func NewModifyTagRetentionRuleResponse() (response *ModifyTagRetentionRuleResponse) {
+    response = &ModifyTagRetentionRuleResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 更新版本保留规则
+func (c *Client) ModifyTagRetentionRule(request *ModifyTagRetentionRuleRequest) (response *ModifyTagRetentionRuleResponse, err error) {
+    if request == nil {
+        request = NewModifyTagRetentionRuleRequest()
+    }
+    response = NewModifyTagRetentionRuleResponse()
     err = c.Send(request, response)
     return
 }